+++ src/DrWatson.jl
@@ -64,7 +64,8 @@ if they two values coincide. I.e. no longer is `1.0` output as `1` in `savename`
 In this new major release, the following breaking changes have occured:
 1. DrWatson now uses, and suggests using, JLD2.jl instead of BSON.jl
    for saving files.
-2. The `savename` stuff.
+2. The behavior of `savename` with respect to rounding has changed,
+   see its docstring for more.
 \n
 """; color = :light_magenta)
 touch(joinpath(@__DIR__, update_name))