+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 * `tmpsave` now saves as JLD2.jl and requires loading explicitly the `"params"` field.
 * `collect_results` saves as JLD2.jl by default now.
 * `produce_or_load` saves as JLD2.jl by default now.
-
-# 1.19.0
 * `savename` will no longer replace a floating point number with its integer version, if they coincide with respect to rounding. Thus, "integer" `AbstractFloat`s will always end with `.0` in `savename`.
+* `savename`'s `scientific` keyword has been replaced by `sigdigits`, as this was the only thing it was doing.
+
 # 1.18.3
 * Remove type constraints on `produce_or_load` path argument (#229)
 # 1.18.2